+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Contributors (alphabetical): Jo Devriendt, Matthias van der hallen, Pierre Carbo
 * (breaking) Merge 101: simplify `get_range` API
 * Merge !103: support alternative "batch" propagation algorithm (not faster)
 * Merge !105: "directional propagation" to improve performance for a sequence of assert+propagate
+* Merge !106: improve performance of Problem.simplify()
 * fix #104: implications, equivalences are binary connectives, not n-ary.
 
 ### Interactive Constructors